[IR] split op_result from value
