+++ paddle/fluid/pir/dialect/operator/ir/manual_api.h
@@ -18,7 +18,7 @@
 
 #include "paddle/phi/common/data_type.h"
 #include "paddle/phi/common/place.h"
-#include "paddle/pir/core/value.h"
+#include "paddle/pir/core/op_result.h"
 
 namespace paddle {
 namespace dialect {
